refactor(tests): Split engine/tests.rs into focused submodules (#230 PR review)
Addresses PRRT...PrQ from the round-3 PR review: src/evaluator/engine/tests.rs grew to 3666 lines, well past the 500-600 line project guideline. Split into the three meta-type suites the reviewer called out.

Structure:

- src/evaluator/engine/tests/mod.rs (2777 lines): core evaluator tests (byte/short/long/string/pstring/relative-offset/recursion) plus centralized meta-type test helpers (make_context_with_env, use_rule, use_rule_at, build_name_table, default_rule, clear_rule, byte_eq_rule, indirect_rule, offset_rule) marked pub(super) so submodules can share them.

- src/evaluator/engine/tests/meta_use_tests.rs (405 lines): MetaType::Use dispatch, subroutine name-table resolution, and subroutine base_offset biasing tests.

- src/evaluator/engine/tests/meta_default_clear_indirect_tests.rs (275 lines): MetaType::Default fallback, MetaType::Clear state-reset, and MetaType::Indirect root re-entry tests.

- src/evaluator/engine/tests/meta_offset_tests.rs (247 lines): MetaType::Offset dispatch and evaluate_children_or_warn helper coverage.

No test content changed -- each test's body is byte-identical to the pre-split version. The three new submodules each have their own SPDX/copyright header and module doc. Test count unchanged (1364/1364 pass, 5 pre-existing skipped).
