+++ src/evaluator/engine/tests/mod.rs
@@ -2632,7 +2632,7 @@ fn test_search_parent_relative_child_at_positive_offset() {
 }
 
 // =============================================================================
-// Tests for MetaType::Use semantics
+// Meta-type test helpers (shared across meta_use_tests, meta_default_clear_indirect_tests, meta_offset_tests)
 // =============================================================================
 
 use crate::evaluator::RuleEnvironment;
@@ -2643,7 +2643,10 @@ use crate::parser::name_table::NameTable;
 /// root-rules list. The root-rules list is retained for parity with the
 /// `RuleEnvironment` shape even though `MetaType::Use` itself does not
 /// consult it.
-fn make_context_with_env(name_table: NameTable, root_rules: &[MagicRule]) -> EvaluationContext {
+pub(super) fn make_context_with_env(
+    name_table: NameTable,
+    root_rules: &[MagicRule],
+) -> EvaluationContext {
     let env = std::sync::Arc::new(RuleEnvironment {
         name_table: std::sync::Arc::new(name_table),
         root_rules: std::sync::Arc::from(root_rules),
@@ -2653,14 +2656,14 @@ fn make_context_with_env(name_table: NameTable, root_rules: &[MagicRule]) -> Eva
 
 /// Minimal helper: wrap a `TypeKind::Meta(MetaType::Use(name))` rule at
 /// offset 0 with the given `message` and empty child list.
-fn use_rule(name: &str) -> MagicRule {
+pub(super) fn use_rule(name: &str) -> MagicRule {
     use_rule_at(name, 0)
 }
 
 /// Build a `Use` rule at a specific use-site offset. Used by tests
 /// that need to prove subroutine `base_offset` biasing actually
 /// depends on the use-site value.
-fn use_rule_at(name: &str, offset: i64) -> MagicRule {
+pub(super) fn use_rule_at(name: &str, offset: i64) -> MagicRule {
     MagicRule {
         offset: OffsetSpec::Absolute(offset),
         typ: TypeKind::Meta(MetaType::Use(name.to_string())),
@@ -2674,7 +2677,7 @@ fn use_rule_at(name: &str, offset: i64) -> MagicRule {
 }
 
 /// Construct a name table from `(name, subroutine_rules)` pairs.
-fn build_name_table(entries: Vec<(&str, Vec<MagicRule>)>) -> NameTable {
+pub(super) fn build_name_table(entries: Vec<(&str, Vec<MagicRule>)>) -> NameTable {
     // Build via the extraction helper so the table construction matches the
     // real parser path. Wrap each entry in a Name rule whose `children` are
     // the subroutine body.
@@ -2695,254 +2698,8 @@ fn build_name_table(entries: Vec<(&str, Vec<MagicRule>)>) -> NameTable {
     table
 }
 
-#[test]
-fn test_use_known_name_evaluates_subroutine() {
-    // The subroutine `part2` reads byte 3 and expects 0x42.
-    let subroutine = vec![MagicRule {
-        offset: OffsetSpec::Absolute(3),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0x42),
-        message: "sub-match".to_string(),
-        children: vec![],
-        level: 1,
-        strength_modifier: None,
-    }];
-    let table = build_name_table(vec![("part2", subroutine)]);
-    let mut context = make_context_with_env(table, &[]);
-
-    let buffer = [0x00u8, 0x00, 0x00, 0x42, 0x00];
-    let rules = vec![use_rule("part2")];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert_eq!(
-        matches.len(),
-        1,
-        "subroutine should produce exactly one match"
-    );
-    assert_eq!(matches[0].message, "sub-match");
-}
-
-#[test]
-fn test_use_unknown_name_returns_no_match() {
-    // Empty name table so the lookup fails; the evaluator should not panic
-    // and should produce zero matches.
-    let table = NameTable::empty();
-    let mut context = make_context_with_env(table, &[]);
-
-    let buffer = [0x00u8, 0x42];
-    let rules = vec![use_rule("nonexistent")];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert!(matches.is_empty(), "unknown name should yield no matches");
-}
-
-#[test]
-fn test_use_without_rule_env_returns_no_match() {
-    // A default context has no rule_env attached; `use` rules should be
-    // silent no-ops in that case rather than returning an error or panicking.
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let buffer = [0x00u8, 0x42];
-    let rules = vec![use_rule("part2")];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert!(
-        matches.is_empty(),
-        "Use with no rule_env should produce no matches"
-    );
-}
-
-#[test]
-fn test_use_recursion_limit() {
-    // Build a mutually-recursive pair: subroutine A calls B, B calls A.
-    // With the default recursion limit, this should surface as
-    // `RecursionLimitExceeded` rather than a stack overflow.
-    let a_body = vec![use_rule("b")];
-    let b_body = vec![use_rule("a")];
-    let table = build_name_table(vec![("a", a_body), ("b", b_body)]);
-    let mut context = make_context_with_env(table, &[]);
-
-    let buffer = [0u8; 8];
-    let rules = vec![use_rule("a")];
-    let result = evaluate_rules(&rules, &buffer, &mut context);
-    assert!(
-        matches!(
-            result,
-            Err(LibmagicError::EvaluationError(
-                crate::error::EvaluationError::RecursionLimitExceeded { .. }
-            ))
-        ),
-        "mutual recursion through use must surface RecursionLimitExceeded, got {result:?}"
-    );
-}
-
-#[test]
-fn test_use_child_rules_evaluated_after_subroutine() {
-    // `Use` itself does not expose a visible RuleMatch today, so we cover
-    // the "subroutine matches come first" invariant by verifying that the
-    // subroutine's match appears in the output and is followed by a
-    // sibling rule's match in the surrounding scope.
-    //
-    // `EvaluationConfig::default()` sets `stop_at_first_match = true`, which
-    // (correctly, after the Comment 2 fix) short-circuits sibling iteration
-    // once the `use` path produces a match. To exercise the ordering
-    // invariant between the subroutine and its sibling we opt into the
-    // "completeness" semantics by disabling first-match short-circuit.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let subroutine = vec![MagicRule {
-        offset: OffsetSpec::Absolute(0),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0xAA),
-        message: "sub-head".to_string(),
-        children: vec![],
-        level: 1,
-        strength_modifier: None,
-    }];
-    let table = build_name_table(vec![("sub", subroutine)]);
-    let env = std::sync::Arc::new(RuleEnvironment {
-        name_table: std::sync::Arc::new(table),
-        root_rules: std::sync::Arc::from(&[] as &[MagicRule]),
-    });
-    let mut context = EvaluationContext::new(config).with_rule_env(env);
-
-    let buffer = [0xAAu8, 0xBB, 0xCC];
-    let sibling = MagicRule {
-        offset: OffsetSpec::Absolute(1),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0xBB),
-        message: "sibling".to_string(),
-        children: vec![],
-        level: 0,
-        strength_modifier: None,
-    };
-    let rules = vec![use_rule("sub"), sibling];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert_eq!(matches.len(), 2);
-    assert_eq!(matches[0].message, "sub-head");
-    assert_eq!(matches[1].message, "sibling");
-}
-
-#[test]
-fn test_use_stop_at_first_match_short_circuits_siblings() {
-    // Comment 2 regression guard: with the default
-    // `stop_at_first_match = true` config, a successful `use` subroutine
-    // must prevent later sibling top-level rules from being evaluated,
-    // matching the short-circuit semantics every other rule kind obeys.
-    let subroutine = vec![MagicRule {
-        offset: OffsetSpec::Absolute(0),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0xAA),
-        message: "sub-head".to_string(),
-        children: vec![],
-        level: 1,
-        strength_modifier: None,
-    }];
-    let table = build_name_table(vec![("sub", subroutine)]);
-    let mut context = make_context_with_env(table, &[]);
-
-    let buffer = [0xAAu8, 0xBB, 0xCC];
-    let sibling = MagicRule {
-        offset: OffsetSpec::Absolute(1),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0xBB),
-        message: "sibling".to_string(),
-        children: vec![],
-        level: 0,
-        strength_modifier: None,
-    };
-    let rules = vec![use_rule("sub"), sibling];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert_eq!(
-        matches.len(),
-        1,
-        "stop-at-first-match must halt sibling iteration once the use path produces a match"
-    );
-    assert_eq!(matches[0].message, "sub-head");
-}
-
-#[test]
-fn test_use_rule_children_are_evaluated() {
-    // Comment 1 regression guard: a `use` rule with its own children must
-    // descend into those children after the subroutine runs, so that
-    // libmagic chains like `>>0 use part2` followed by continuation rules
-    // continue producing matches in document order.
-    let subroutine = vec![MagicRule {
-        offset: OffsetSpec::Absolute(0),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0xAA),
-        message: "sub-head".to_string(),
-        children: vec![],
-        level: 1,
-        strength_modifier: None,
-    }];
-    let table = build_name_table(vec![("sub", subroutine)]);
-    // Disable stop-at-first-match so both the subroutine and the child
-    // rule are visible in the match vector.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let env = std::sync::Arc::new(RuleEnvironment {
-        name_table: std::sync::Arc::new(table),
-        root_rules: std::sync::Arc::from(&[] as &[MagicRule]),
-    });
-    let mut context = EvaluationContext::new(config).with_rule_env(env);
-
-    let child = MagicRule {
-        offset: OffsetSpec::Absolute(1),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0xBB),
-        message: "use-child".to_string(),
-        children: vec![],
-        level: 1,
-        strength_modifier: None,
-    };
-    let mut use_with_child = use_rule("sub");
-    use_with_child.children = vec![child];
-
-    let buffer = [0xAAu8, 0xBB, 0xCC];
-    let matches = evaluate_rules(&[use_with_child], &buffer, &mut context).unwrap();
-    assert_eq!(
-        matches.len(),
-        2,
-        "use rule's own children must run after the subroutine"
-    );
-    assert_eq!(matches[0].message, "sub-head");
-    assert_eq!(matches[1].message, "use-child");
-}
-
-#[test]
-fn test_name_rule_leaked_is_noop() {
-    // Programmatic consumers may construct a Name rule directly and pass
-    // it to the evaluator (e.g. property tests). The evaluator must not
-    // panic; it should instead treat the rule as a silent no-op.
-    let leaked = MagicRule {
-        offset: OffsetSpec::Absolute(0),
-        typ: TypeKind::Meta(MetaType::Name("orphan".to_string())),
-        op: Operator::Equal,
-        value: Value::Uint(0),
-        message: String::new(),
-        children: vec![],
-        level: 0,
-        strength_modifier: None,
-    };
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let matches = evaluate_rules(&[leaked], &[0u8; 4], &mut context).unwrap();
-    assert!(matches.is_empty(), "leaked Name rule should be a no-op");
-}
-
-// =============================================================================
-// MetaType::Default / Clear / Indirect tests
-// =============================================================================
-
 /// Build a `Default` rule with the given message and (optional) children.
-fn default_rule(message: &str, children: Vec<MagicRule>) -> MagicRule {
+pub(super) fn default_rule(message: &str, children: Vec<MagicRule>) -> MagicRule {
     MagicRule {
         offset: OffsetSpec::Absolute(0),
         typ: TypeKind::Meta(MetaType::Default),
@@ -2957,7 +2714,7 @@ fn default_rule(message: &str, children: Vec<MagicRule>) -> MagicRule {
 
 /// Build a `Clear` rule. Carries no message in the magic file syntax, but the
 /// AST requires a message field.
-fn clear_rule() -> MagicRule {
+pub(super) fn clear_rule() -> MagicRule {
     MagicRule {
         offset: OffsetSpec::Absolute(0),
         typ: TypeKind::Meta(MetaType::Clear),
@@ -2971,7 +2728,7 @@ fn clear_rule() -> MagicRule {
 }
 
 /// Build a single byte-equality rule at `offset` for `value`.
-fn byte_eq_rule(offset: i64, value: u64, message: &str) -> MagicRule {
+pub(super) fn byte_eq_rule(offset: i64, value: u64, message: &str) -> MagicRule {
     MagicRule {
         offset: OffsetSpec::Absolute(offset),
         typ: TypeKind::Byte { signed: false },
@@ -2984,173 +2741,8 @@ fn byte_eq_rule(offset: i64, value: u64, message: &str) -> MagicRule {
     }
 }
 
-#[test]
-fn test_default_fires_when_no_sibling_matched() {
-    let rules = vec![default_rule("DEFAULT-FIRES", vec![])];
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let matches = evaluate_rules(&rules, &[0u8; 4], &mut context).unwrap();
-    assert_eq!(
-        matches.len(),
-        1,
-        "default with no prior sibling match should fire"
-    );
-    assert_eq!(matches[0].message, "DEFAULT-FIRES");
-}
-
-#[test]
-fn test_default_skipped_when_sibling_matched() {
-    // Disable stop-at-first-match so we can see whether the default would
-    // have fired or not.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let rules = vec![
-        byte_eq_rule(0, 0xAA, "real-match"),
-        default_rule("DEFAULT-SKIPPED", vec![]),
-    ];
-    let mut context = EvaluationContext::new(config);
-    let buffer = [0xAAu8, 0xBB];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert_eq!(
-        matches.len(),
-        1,
-        "default after a successful sibling should not fire"
-    );
-    assert_eq!(matches[0].message, "real-match");
-}
-
-#[test]
-fn test_default_fires_only_once() {
-    // Two consecutive default rules: the first sets sibling_matched, so
-    // the second must not fire.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let rules = vec![
-        default_rule("FIRST-DEFAULT", vec![]),
-        default_rule("SECOND-DEFAULT", vec![]),
-    ];
-    let mut context = EvaluationContext::new(config);
-    let matches = evaluate_rules(&rules, &[0u8; 4], &mut context).unwrap();
-    assert_eq!(
-        matches.len(),
-        1,
-        "only the first default should fire when no real sibling matched"
-    );
-    assert_eq!(matches[0].message, "FIRST-DEFAULT");
-}
-
-#[test]
-fn test_default_children_evaluated() {
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let child = byte_eq_rule(0, 0xAA, "default-child");
-    let rules = vec![default_rule("PARENT-DEFAULT", vec![child])];
-    let mut context = EvaluationContext::new(config);
-    let buffer = [0xAAu8, 0xBB];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert_eq!(
-        matches.len(),
-        2,
-        "default rule's children must be evaluated when the default fires"
-    );
-    assert_eq!(matches[0].message, "PARENT-DEFAULT");
-    assert_eq!(matches[1].message, "default-child");
-}
-
-#[test]
-fn test_clear_resets_sibling_matched() {
-    // Sequence: byte-match, default-skipped, clear, default-fires.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let rules = vec![
-        byte_eq_rule(0, 0xAA, "byte-match"),
-        default_rule("DEFAULT-SKIPPED", vec![]),
-        clear_rule(),
-        default_rule("DEFAULT-FIRES-AFTER-CLEAR", vec![]),
-    ];
-    let mut context = EvaluationContext::new(config);
-    let buffer = [0xAAu8, 0xBB];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert_eq!(
-        matches.len(),
-        2,
-        "clear must reset sibling_matched so a later default fires"
-    );
-    assert_eq!(matches[0].message, "byte-match");
-    assert_eq!(matches[1].message, "DEFAULT-FIRES-AFTER-CLEAR");
-}
-
-#[test]
-fn test_clear_at_top_is_noop() {
-    let rules = vec![clear_rule(), default_rule("AFTER-CLEAR", vec![])];
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let matches = evaluate_rules(&rules, &[0u8; 4], &mut context).unwrap();
-    assert_eq!(
-        matches.len(),
-        1,
-        "clear at top of list is a no-op; default after still fires"
-    );
-    assert_eq!(matches[0].message, "AFTER-CLEAR");
-}
-
-#[test]
-fn test_clear_does_not_produce_match() {
-    let rules = vec![clear_rule()];
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let matches = evaluate_rules(&rules, &[0u8; 4], &mut context).unwrap();
-    assert!(matches.is_empty(), "clear alone must produce no match");
-}
-
-#[test]
-fn test_default_clear_per_level_isolation() {
-    // Parent has its own sibling_matched flag. The child list runs with a
-    // fresh flag, so a child-level `default` must fire even though the
-    // parent's flag is true.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let parent = MagicRule {
-        offset: OffsetSpec::Absolute(0),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0xAA),
-        message: "parent-match".to_string(),
-        children: vec![
-            byte_eq_rule(1, 0xBB, "child-byte-match"),
-            default_rule("CHILD-DEFAULT-SKIPPED", vec![]),
-            clear_rule(),
-            default_rule("CHILD-DEFAULT-AFTER-CLEAR", vec![]),
-        ],
-        level: 0,
-        strength_modifier: None,
-    };
-    let mut context = EvaluationContext::new(config);
-    let buffer = [0xAAu8, 0xBB];
-    let matches = evaluate_rules(&[parent], &buffer, &mut context).unwrap();
-
-    // Expected order: parent-match, child-byte-match, CHILD-DEFAULT-AFTER-CLEAR
-    let messages: Vec<&str> = matches.iter().map(|m| m.message.as_str()).collect();
-    assert_eq!(
-        messages,
-        vec![
-            "parent-match",
-            "child-byte-match",
-            "CHILD-DEFAULT-AFTER-CLEAR"
-        ],
-        "child-level sibling_matched must be isolated from parent-level state"
-    );
-}
-
 /// Build an `Indirect` rule at `offset` with optional children.
-fn indirect_rule(offset: i64, message: &str, children: Vec<MagicRule>) -> MagicRule {
+pub(super) fn indirect_rule(offset: i64, message: &str, children: Vec<MagicRule>) -> MagicRule {
     MagicRule {
         offset: OffsetSpec::Absolute(offset),
         typ: TypeKind::Meta(MetaType::Indirect),
@@ -3163,111 +2755,9 @@ fn indirect_rule(offset: i64, message: &str, children: Vec<MagicRule>) -> MagicR
     }
 }
 
-#[test]
-fn test_indirect_evaluates_root_rules_at_offset() {
-    // Root rules: detect a "ZIP-like" header (0x50 0x4b) at offset 0 of the
-    // sub-buffer. The indirect rule fires at offset 4 of the outer buffer,
-    // which means the sub-buffer starts at byte 4. Place 0x50 0x4b there.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let root_rule = byte_eq_rule(0, 0x50, "ZIP-like-header");
-    let root_rules: Vec<MagicRule> = vec![root_rule];
-
-    // Build an environment where root_rules is the same as the rules we
-    // dispatch into.
-    let env = std::sync::Arc::new(RuleEnvironment {
-        name_table: std::sync::Arc::new(NameTable::empty()),
-        root_rules: std::sync::Arc::from(root_rules.as_slice()),
-    });
-    let mut context = EvaluationContext::new(config).with_rule_env(env);
-
-    // Buffer: ELF magic at offset 0, ZIP-like at offset 4. The indirect
-    // rule is the trigger; the root re-entry detects 0x50 at sub-buffer 0.
-    let buffer = [0x7fu8, 0x45, 0x4c, 0x46, 0x50, 0x4b, 0x03, 0x04];
-    let rules = vec![indirect_rule(4, "indirect-trigger", vec![])];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-
-    assert!(
-        matches.iter().any(|m| m.message == "ZIP-like-header"),
-        "indirect must dispatch root rules against the sub-buffer at offset 4; got {matches:?}"
-    );
-}
-
-#[test]
-fn test_indirect_out_of_bounds_is_noop() {
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let env = std::sync::Arc::new(RuleEnvironment {
-        name_table: std::sync::Arc::new(NameTable::empty()),
-        root_rules: std::sync::Arc::from(&[byte_eq_rule(0, 0x00, "root")] as &[MagicRule]),
-    });
-    let mut context = EvaluationContext::new(config).with_rule_env(env);
-
-    let buffer = [0u8; 4];
-    // Indirect at offset 100, which is well past the 4-byte buffer.
-    let rules = vec![indirect_rule(100, "indirect-oob", vec![])];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert!(
-        matches.is_empty(),
-        "indirect past buffer end must be a graceful no-op"
-    );
-}
-
-#[test]
-fn test_indirect_without_env_is_noop() {
-    // Property tests synthesize Indirect rules without an attached
-    // RuleEnvironment, so this path must be a graceful no-op (matching the
-    // `Use`-without-env contract). The engine logs at `debug!` rather than
-    // panicking via `debug_assert!` to preserve the never-panics invariant
-    // exercised by `prop_arbitrary_rule_evaluation_never_panics`.
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let buffer = [0u8; 4];
-    let rules = vec![indirect_rule(0, "indirect-no-env", vec![])];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert!(
-        matches.is_empty(),
-        "indirect without env must produce no matches"
-    );
-}
-
-#[test]
-fn test_indirect_recursion_limit() {
-    // Root rules contain an indirect rule that points back to offset 0,
-    // creating an infinite re-entry chain. Must surface as
-    // `RecursionLimitExceeded`, not stack overflow.
-    let inner_indirect = indirect_rule(0, "recursive-indirect", vec![]);
-    let root_rules: Vec<MagicRule> = vec![inner_indirect];
-    let env = std::sync::Arc::new(RuleEnvironment {
-        name_table: std::sync::Arc::new(NameTable::empty()),
-        root_rules: std::sync::Arc::from(root_rules.as_slice()),
-    });
-    let mut context = EvaluationContext::new(EvaluationConfig::default()).with_rule_env(env);
-
-    let buffer = [0u8; 8];
-    let rules = vec![indirect_rule(0, "outer-indirect", vec![])];
-    let result = evaluate_rules(&rules, &buffer, &mut context);
-    assert!(
-        matches!(
-            result,
-            Err(LibmagicError::EvaluationError(
-                crate::error::EvaluationError::RecursionLimitExceeded { .. }
-            ))
-        ),
-        "infinite indirect recursion must surface RecursionLimitExceeded, got {result:?}"
-    );
-}
-
-// =======================================================================
-// MetaType::Offset dispatch (issue #42)
-// =======================================================================
-
 /// Build an `Offset` rule at `offset` with an `x` (`AnyValue`) operator and
 /// the given message. Mirrors `default_rule`/`indirect_rule` helpers.
-fn offset_rule(offset: i64, message: &str, children: Vec<MagicRule>) -> MagicRule {
+pub(super) fn offset_rule(offset: i64, message: &str, children: Vec<MagicRule>) -> MagicRule {
     MagicRule {
         offset: OffsetSpec::Absolute(offset),
         typ: TypeKind::Meta(MetaType::Offset),
@@ -3280,387 +2770,10 @@ fn offset_rule(offset: i64, message: &str, children: Vec<MagicRule>) -> MagicRul
     }
 }
 
-#[test]
-fn test_offset_emits_match_with_resolved_position() {
-    let rules = vec![offset_rule(5, "pos=%lld", vec![])];
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let matches = evaluate_rules(&rules, &[0u8; 10], &mut context).unwrap();
-    assert_eq!(matches.len(), 1, "offset rule must emit exactly one match");
-    assert_eq!(matches[0].offset, 5, "match.offset is the resolved offset");
-    assert_eq!(
-        matches[0].value,
-        Value::Uint(5),
-        "match.value carries the resolved offset for format substitution"
-    );
-    assert_eq!(matches[0].message, "pos=%lld");
-}
-
-#[test]
-fn test_offset_at_zero() {
-    // Regression guard: offset 0 must still produce a match (not be
-    // indistinguishable from "no match").
-    let rules = vec![offset_rule(0, "top", vec![])];
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let matches = evaluate_rules(&rules, &[0u8; 4], &mut context).unwrap();
-    assert_eq!(matches.len(), 1);
-    assert_eq!(matches[0].value, Value::Uint(0));
-}
-
-#[test]
-fn test_offset_out_of_bounds_graceful_skip() {
-    // Offset past the end of the buffer is a data-dependent skip, not an
-    // error. Matches the Indirect dispatch's graceful-skip discipline.
-    let rules = vec![offset_rule(1_000_000, "unreachable", vec![])];
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let matches = evaluate_rules(&rules, &[0u8; 4], &mut context).unwrap();
-    assert!(
-        matches.is_empty(),
-        "offset past buffer end must produce no match"
-    );
-}
-
-#[test]
-fn test_offset_non_x_operator_is_skipped() {
-    // magic(5) only allows `x` on an `offset` rule. Anything else is
-    // semantically undefined -> debug-log + skip.
-    let mut rule = offset_rule(0, "bogus", vec![]);
-    rule.op = Operator::Equal;
-    rule.value = Value::Uint(5);
-    let rules = vec![rule];
-    let mut context = EvaluationContext::new(EvaluationConfig::default());
-    let matches = evaluate_rules(&rules, &[0u8; 4], &mut context).unwrap();
-    assert!(
-        matches.is_empty(),
-        "offset rule with non-AnyValue operator must be skipped"
-    );
-}
-
-#[test]
-fn test_offset_evaluates_children() {
-    // A child byte rule at offset 0 runs AFTER the parent offset rule
-    // fires. The child's own offset is resolved independently.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let mut parent = offset_rule(
-        0,
-        "parent-offset",
-        vec![byte_eq_rule(0, 0x42, "child-byte")],
-    );
-    // Child level must be deeper than parent per MagicRule::validate.
-    parent.children[0].level = 1;
-    let buffer = [0x42u8, 0x00, 0x00];
-    let mut context = EvaluationContext::new(config);
-    let matches = evaluate_rules(&[parent], &buffer, &mut context).unwrap();
-    let messages: Vec<&str> = matches.iter().map(|m| m.message.as_str()).collect();
-    assert_eq!(messages, vec!["parent-offset", "child-byte"]);
-}
-
-#[test]
-fn test_offset_advances_anchor_for_children() {
-    // An offset rule at position 5 advances `last_match_end` to 5 *for its
-    // children* -- but NOT for sibling rules at the same level. This
-    // matches libmagic's continuation-level semantics: each sibling at
-    // level L resolves `&N` against the parent-level anchor, not against
-    // the previous sibling's advance. See the `entry_anchor` discipline
-    // in `evaluate_rules`.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-
-    // A child of the offset rule uses &0 to resolve at the offset rule's
-    // resolved position (5). buffer[5] = 0x42.
-    let mut child = byte_eq_rule(0, 0x42, "child-at-offset-anchor");
-    child.offset = OffsetSpec::Relative(0);
-    child.level = 1;
-
-    let buffer = [0x00u8, 0x00, 0x00, 0x00, 0x00, 0x42, 0x00];
-    let rules = vec![offset_rule(5, "mark", vec![child])];
-    let mut context = EvaluationContext::new(config);
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-
-    assert!(
-        matches
-            .iter()
-            .any(|m| m.message == "child-at-offset-anchor"),
-        "child of offset rule must resolve against offset's anchor (5); got {matches:?}"
-    );
-}
-
-#[test]
-fn test_offset_does_not_advance_anchor_for_continuation_siblings() {
-    // Regression guard for the libmagic continuation-sibling anchor
-    // semantic: two CHILD siblings at the same level resolve `&N`
-    // against the parent-level anchor, not against the previous
-    // sibling's advance. This is gated on `recursion_depth > 0`;
-    // top-level siblings still chain (see
-    // `relative_anchor_can_decrease_...` in the relative-offset
-    // integration tests).
-    //
-    // Parent `byte` at offset 0 matches 0x01 -> anchor = 1. Two
-    // child siblings at &0 must both read buffer[1] = 0x42. If the
-    // first child incorrectly advanced the anchor to 2, the second
-    // would read buffer[2] = 0x00 and miss.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let parent = MagicRule {
-        offset: OffsetSpec::Absolute(0),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0x01),
-        message: "parent".to_string(),
-        children: vec![
-            MagicRule {
-                offset: OffsetSpec::Relative(0),
-                typ: TypeKind::Byte { signed: false },
-                op: Operator::Equal,
-                value: Value::Uint(0x42),
-                message: "sibling-1".to_string(),
-                children: vec![],
-                level: 1,
-                strength_modifier: None,
-            },
-            MagicRule {
-                offset: OffsetSpec::Relative(0),
-                typ: TypeKind::Byte { signed: false },
-                op: Operator::Equal,
-                value: Value::Uint(0x42),
-                message: "sibling-2".to_string(),
-                children: vec![],
-                level: 1,
-                strength_modifier: None,
-            },
-        ],
-        level: 0,
-        strength_modifier: None,
-    };
-
-    let buffer = [0x01u8, 0x42, 0x00, 0x00];
-    let mut context = EvaluationContext::new(config);
-    let matches = evaluate_rules(&[parent], &buffer, &mut context).unwrap();
-    let messages: Vec<&str> = matches.iter().map(|m| m.message.as_str()).collect();
-    assert_eq!(
-        messages,
-        vec!["parent", "sibling-1", "sibling-2"],
-        "both continuation siblings must resolve against parent anchor (1); \
-         if sibling-1 advanced the anchor to 2, sibling-2 would read \
-         buffer[2]=0x00 and fail"
-    );
-}
-
-#[test]
-fn test_offset_sets_sibling_matched() {
-    // An offset rule match suppresses a following `default` sibling --
-    // same discipline as any other matching rule.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-    let rules = vec![
-        offset_rule(0, "offset-match", vec![]),
-        default_rule("DEFAULT-SUPPRESSED", vec![]),
-    ];
-    let mut context = EvaluationContext::new(config);
-    let matches = evaluate_rules(&rules, &[0u8; 4], &mut context).unwrap();
-    let messages: Vec<&str> = matches.iter().map(|m| m.message.as_str()).collect();
-    assert_eq!(
-        messages,
-        vec!["offset-match"],
-        "default must be suppressed when offset sibling matched; got {matches:?}"
-    );
-}
-
-// =======================================================================
-// Subroutine base_offset biasing (issue #42 -- use-site offset
-// propagation). Critical coverage per post-PR code review.
-// =======================================================================
-
-#[test]
-fn test_use_subroutine_absolute_offset_biased_by_use_site() {
-    // Regression guard: if `SubroutineScope::enter` fails to seed
-    // `base_offset` with the use-site offset, a subroutine rule at
-    // `Absolute(0)` will read from buffer[0] instead of
-    // buffer[use_site]. This test proves the bias is active by
-    // placing distinct magic bytes at two different positions and
-    // verifying that the subroutine reads the use-site one.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-
-    // Subroutine body: a single rule reading at Absolute(0). Without
-    // base_offset biasing this resolves to file position 0. With
-    // biasing it resolves to the use-site (position 8 in this test).
-    let subroutine_body = vec![byte_eq_rule(0, 0x42, "sub-match-at-base")];
-    let name_table = build_name_table(vec![("sub", subroutine_body)]);
-    let env = std::sync::Arc::new(RuleEnvironment {
-        name_table: std::sync::Arc::new(name_table),
-        root_rules: std::sync::Arc::from(&[] as &[MagicRule]),
-    });
-
-    // Use-site at offset 8. buffer[0] = 0x00 (would fail with bias
-    // missing); buffer[8] = 0x42 (required for bias-active success).
-    let mut buffer = vec![0u8; 16];
-    buffer[8] = 0x42;
-
-    let mut context = EvaluationContext::new(config).with_rule_env(env);
-    let rules = vec![use_rule_at("sub", 8)];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert!(
-        matches.iter().any(|m| m.message == "sub-match-at-base"),
-        "subroutine rule at Absolute(0) must be biased by use-site offset 8 \
-         -- reading buffer[8] = 0x42. If bias missing, reads buffer[0] = 0x00 \
-         and the test fails. got {matches:?}"
-    );
-}
-
-#[test]
-fn test_use_subroutine_relative_offset_unaffected_by_use_site() {
-    // Companion to the bias test above: `Relative(N)` is resolved
-    // against `last_match_end`, which `SubroutineScope` also seeds
-    // to the use-site. We verify the Relative rule reads at the
-    // use-site + N, NOT at use-site + base + N (which would be a
-    // double-bias bug).
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-
-    // Subroutine body: a Relative(0) rule that reads at the
-    // use-site (seeded via last_match_end).
-    let mut rel_rule = byte_eq_rule(0, 0x42, "rel-sub-match");
-    rel_rule.offset = OffsetSpec::Relative(0);
-    let subroutine_body = vec![rel_rule];
-    let name_table = build_name_table(vec![("rsub", subroutine_body)]);
-    let env = std::sync::Arc::new(RuleEnvironment {
-        name_table: std::sync::Arc::new(name_table),
-        root_rules: std::sync::Arc::from(&[] as &[MagicRule]),
-    });
-
-    let mut buffer = vec![0u8; 16];
-    buffer[5] = 0x42;
-
-    let mut context = EvaluationContext::new(config).with_rule_env(env);
-    let rules = vec![use_rule_at("rsub", 5)];
-    let matches = evaluate_rules(&rules, &buffer, &mut context).unwrap();
-    assert!(
-        matches.iter().any(|m| m.message == "rel-sub-match"),
-        "subroutine Relative(0) rule must read at use-site (5) via last_match_end, \
-         not at use-site+base (10). got {matches:?}"
-    );
-}
-
-#[test]
-fn test_continuation_sibling_reset_after_bytes_consumed() {
-    // Stronger regression guard than
-    // `test_offset_does_not_advance_anchor_for_continuation_siblings`,
-    // which used Relative(0) on both siblings and was trivially
-    // non-advancing. Here the first sibling consumes actual bytes,
-    // so if the `is_child_sibling_list` reset is removed the second
-    // sibling would read from a shifted anchor.
-    //
-    // Parent byte at 0 matches 0x01 -> anchor = 1.
-    // Sibling-1: Long at &0 (resolves to 1, reads 4 bytes,
-    //            advances anchor to 5 WITHOUT the reset).
-    // Sibling-2: Byte at &0 (must resolve to 1, not 5).
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-
-    let long_sibling = MagicRule {
-        offset: OffsetSpec::Relative(0),
-        typ: TypeKind::Long {
-            endian: crate::parser::ast::Endianness::Little,
-            signed: false,
-        },
-        op: Operator::Equal,
-        value: Value::Uint(0x0403_0201),
-        message: "long-sibling".to_string(),
-        children: vec![],
-        level: 1,
-        strength_modifier: None,
-    };
-    let byte_sibling = MagicRule {
-        offset: OffsetSpec::Relative(0),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        // buffer[1] = 0x01 -- if reset is removed, sibling-2 reads
-        // buffer[5] instead and matches 0x42 (wrong!).
-        value: Value::Uint(0x01),
-        message: "byte-sibling-sees-parent-anchor".to_string(),
-        children: vec![],
-        level: 1,
-        strength_modifier: None,
-    };
-    let parent = MagicRule {
-        offset: OffsetSpec::Absolute(0),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0x01),
-        message: "parent".to_string(),
-        children: vec![long_sibling, byte_sibling],
-        level: 0,
-        strength_modifier: None,
-    };
-
-    // buffer[0]=0x01 parent; buffer[1..5]=0x01,0x02,0x03,0x04 long
-    // match; buffer[5]=0x42 bait for missing-reset failure.
-    let buffer = [0x01u8, 0x01, 0x02, 0x03, 0x04, 0x42, 0x00];
-    let mut context = EvaluationContext::new(config);
-    let matches = evaluate_rules(&[parent], &buffer, &mut context).unwrap();
-    let messages: Vec<&str> = matches.iter().map(|m| m.message.as_str()).collect();
-    assert_eq!(
-        messages,
-        vec!["parent", "long-sibling", "byte-sibling-sees-parent-anchor"],
-        "byte-sibling must read buffer[1]=0x01 via parent-level anchor reset; \
-         if reset is missing it reads buffer[5]=0x42 and test fails. got {matches:?}"
-    );
-}
-
-// =======================================================================
-// evaluate_children_or_warn graceful-error helper (issue #42 close-out)
-// =======================================================================
-
-#[test]
-fn test_evaluate_children_or_warn_swallows_buffer_overrun_keeps_parent_match() {
-    // Regression guard for the extracted `evaluate_children_or_warn`
-    // helper: a child with an absolute offset past the buffer end must
-    // produce a `BufferOverrun` that is swallowed (warn-logged) rather
-    // than propagated. The parent match must still appear in the
-    // results. Covers the graceful-skip arm for all four dispatch
-    // sites (Default/Indirect/Offset/Use) via the Offset arm -- they
-    // all delegate to the same helper.
-    let config = EvaluationConfig {
-        stop_at_first_match: false,
-        ..EvaluationConfig::default()
-    };
-
-    // Child rule at absolute offset 1000 reads a byte -- far past the
-    // tiny buffer we supply. The helper should catch the BufferOverrun
-    // and warn-log, not fail the evaluation.
-    let child = MagicRule {
-        offset: OffsetSpec::Absolute(1000),
-        typ: TypeKind::Byte { signed: false },
-        op: Operator::Equal,
-        value: Value::Uint(0x00),
-        message: "unreachable-child".to_string(),
-        children: vec![],
-        level: 1,
-        strength_modifier: None,
-    };
-    let parent = offset_rule(0, "parent-offset-match", vec![child]);
-
-    let mut context = EvaluationContext::new(config);
-    let matches = evaluate_rules(&[parent], &[0u8; 4], &mut context).unwrap();
-    let messages: Vec<&str> = matches.iter().map(|m| m.message.as_str()).collect();
-    assert_eq!(
-        messages,
-        vec!["parent-offset-match"],
-        "parent match must survive a child's BufferOverrun; child must be silently skipped, got {matches:?}"
-    );
-}
+// Submodule declarations
+#[cfg(test)]
+mod meta_default_clear_indirect_tests;
+#[cfg(test)]
+mod meta_offset_tests;
+#[cfg(test)]
+mod meta_use_tests;
